refactor(entity-browser): add custom actions to entity-browser
- to support custom actions in modal and not only in fullscreen mode
- add delete and qr-code action support

Refs: TOCEV-2591
