+++ packages/entity-browser/package.json
@@ -9,12 +9,13 @@
   },
   "devDependencies": {
     "tocco-app-extensions": "*",
+    "tocco-delete": "*",
     "tocco-entity-detail": "*",
     "tocco-entity-list": "*",
-    "tocco-input-edit": "*",
     "tocco-test-util": "*",
     "tocco-theme": "*",
     "tocco-ui": "*",
-    "tocco-util": "*"
+    "tocco-util": "*",
+    "tocco-user-qr-action": "*"
   }
 }
