+++ packages/entity-browser/src/components/LazyAction/actions/Delete.js
@@ -1,0 +1,1 @@
+export {default} from 'tocco-delete/src/main'
