+++ packages/entity-browser/src/components/LazyAction/actions/UserQrAction.js
@@ -1,0 +1,1 @@
+export {default} from 'tocco-user-qr-action/src/main'
