+++ packages/entity-browser/src/components/LazyAction/index.js
@@ -1,0 +1,1 @@
+export {default} from './LazyAction'
