+++ packages/entity-browser/src/routes/action/components/ActionView/index.js
@@ -1,0 +1,1 @@
+export {default} from './ActionViewContainer'
