+++ packages/entity-browser/src/routes/action/index.js
@@ -1,0 +1,5 @@
+import {route} from 'tocco-util'
+
+export default (store, input) => {
+  return route.loadRoute(store, input, () => (import('./route')))
+}
