+++ packages/entity-browser/src/routes/action/route.js
@@ -1,0 +1,8 @@
+import ActionView from './components/ActionView'
+
+export default {
+  container: ActionView,
+  reducers: {},
+  sagas: [],
+  inputDispatches: []
+}
