--- packages/entity-browser/src/routes/app/route.js
@@ -1,8 +1,0 @@
-import LazyAction from './containers/LazyAction'
-
-export default {
-  container: LazyAction,
-  reducers: {},
-  sagas: [],
-  inputDispatches: []
-}
